server/requirements.txt
fastapi
uvicorn

sqlalchemy
psycopg2-binary
alembic

python-dotenv
pydantic
pydantic-settings

langchain
langgraph
langchain-google-genai

tenacity
httpx

cloudinary
geopy
tweepy

transformers
torch
Pillow

passlib[bcrypt]
python-jose[pycryptodome]
argon2_cffi

python-socketio